+++ CHANGELOG.md
@@ -17,6 +17,7 @@
  * Added API function `get_telemetry_data` to the vehicle actor.
  * PythonAPI `Sensor.is_listening` was defined twice (property and method), cleaned and clarified it as a method.
  * Added V2X sensors for cooperative awareness message and custom user-defined messages to support vehicle-to-vehicle communication
+ * Added named tuples for BasicAgent.py's detection result to allow for type-hints and better semantics.
 
 
 ## CARLA 0.9.15